refactor(pricing): remove A/B experiment logic and consolidate pricing components
