+++ apps/www/app/pricing/PricingContent.tsx
@@ -4,20 +4,13 @@ import { ArrowDownIcon } from '@heroicons/react/outline'
 import DefaultLayout from '~/components/Layouts/Default'
 import PricingPlans from '~/components/Pricing/PricingPlans'
 import { useOrganizations } from '~/data/organizations'
-import { hasConsented, posthogClient } from 'common'
 import { ArrowUpRight } from 'lucide-react'
 import dynamic from 'next/dynamic'
 import { usePathname } from 'next/navigation'
-import { useEffect, useState } from 'react'
+import { useEffect } from 'react'
 import { Button } from 'ui'
 import { InfoTooltip } from 'ui-patterns/info-tooltip'
 
-const EXPERIMENT_ID = 'pricingCalculatorExperiment' as const
-type PricingCalculatorVariant = 'control' | 'test'
-
-const NewPricingComputeSection = dynamic(
-  () => import('~/components/Pricing/NewPricingComputeSection')
-)
 const PricingComputeSection = dynamic(() => import('~/components/Pricing/PricingComputeSection'))
 const PricingAddons = dynamic(() => import('~/components/Pricing/PricingAddons'))
 const PricingComparisonTable = dynamic(() => import('~/components/Pricing/PricingComparisonTable'))
@@ -51,30 +44,6 @@ export default function PricingContent() {
 
   const { isLoading, organizations } = useOrganizations()
   const hasExistingOrganizations = !isLoading && organizations.length > 0
-
-  // Pricing calculator A/B experiment
-  // Uses client-side PostHog directly — server-side evaluation lacks full person context for www pages.
-  // DevToolbar overrides (x-ph-flag-overrides cookie) are respected in local dev via posthogClient.getFeatureFlag.
-  const [flagValue, setFlagValue] = useState<PricingCalculatorVariant | false | undefined>(
-    () =>
-      posthogClient.getFeatureFlag(EXPERIMENT_ID) as PricingCalculatorVariant | false | undefined
-  )
-
-  useEffect(() => {
-    return posthogClient.onFeatureFlags(() => {
-      const value = posthogClient.getFeatureFlag(EXPERIMENT_ID)
-      setFlagValue(value as PricingCalculatorVariant | false | undefined)
-    })
-  }, [])
-
-  const isTestVariant = flagValue === 'test'
-  const isInExperiment = flagValue === 'control' || flagValue === 'test'
-
-  useEffect(() => {
-    if (!isInExperiment) return
-
-    posthogClient.captureExperimentExposure(EXPERIMENT_ID, { variant: flagValue }, hasConsented())
-  }, [isInExperiment, flagValue])
 
   return (
     <DefaultLayout>
@@ -119,27 +88,25 @@ export default function PricingContent() {
         id="addon-compute"
         className="container relative mx-auto px-4 lg:px-12 pt-16 md:pt-24 lg:pt-32 lg:pb-16"
       >
-        {isTestVariant && (
-          <div className="text-center mb-8 lg:mb-16">
-            <h2 className="text-foreground text-3xl" id="how-compute-pricing-works">
-              How compute pricing works
-            </h2>
-            <p className="text-foreground-light mt-4 text-lg mb-4">
-              Choose a plan, add projects, and see your total cost
-            </p>
-            <div className="flex items-center justify-center gap-1">
-              <span className="py-1 px-3 bg-surface-100 flex items-center gap-1 border rounded-full text-xs text-foreground-lighter">
-                What is &ldquo;compute&rdquo;?
-                <InfoTooltip side="bottom" className="max-w-[280px]">
-                  Think of compute as the computer your database runs on. As your app grows, you
-                  scale CPU and memory to handle more traffic and data.
-                </InfoTooltip>
-              </span>
-            </div>
+        <div className="text-center mb-8 lg:mb-16">
+          <h2 className="text-foreground text-3xl" id="how-compute-pricing-works">
+            How compute pricing works
+          </h2>
+          <p className="text-foreground-light mt-4 text-lg mb-4">
+            Choose a plan, add projects, and see your total cost
+          </p>
+          <div className="flex items-center justify-center gap-1">
+            <span className="py-1 px-3 bg-surface-100 flex items-center gap-1 border rounded-full text-xs text-foreground-lighter">
+              What is &ldquo;compute&rdquo;?
+              <InfoTooltip side="bottom" className="max-w-[280px]">
+                Think of compute as the computer your database runs on. As your app grows, you scale
+                CPU and memory to handle more traffic and data.
+              </InfoTooltip>
+            </span>
           </div>
-        )}
+        </div>
 
-        {isTestVariant ? <NewPricingComputeSection /> : <PricingComputeSection />}
+        <PricingComputeSection />
       </div>
 
       <div
