+++ packages/common/telemetry-constants.ts
@@ -3019,27 +3019,6 @@ export interface RlsEventTriggerBannerCreateButtonClickedEvent {
 }
 
 /**
- * User was exposed to the pricing calculator experiment on the /pricing page.
- *
- * @group Events
- * @source www
- * @page /pricing
- */
-export interface PricingCalculatorExperimentExposedEvent {
-  action: 'pricing_calculator_experiment_exposed'
-  properties: {
-    /**
-     * Experiment identifier for tracking
-     */
-    experiment_id: 'pricingCalculatorExperiment'
-    /**
-     * Experiment variant: 'control' (existing compute section) or 'test' (new compute section)
-     */
-    variant: 'control' | 'test'
-  }
-}
-
-/**
  * User clicked the Run button in the log explorer.
  *
  * @group Events
@@ -3224,4 +3203,3 @@ export type TelemetryEvent =
   | OrgSubmenuOpenedEvent
   | OrgMenuBackClickedEvent
   | OrgMenuItemClickedEvent
-  | PricingCalculatorExperimentExposedEvent
